crawler/requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
cffi==1.11.5
constantly==15.1.0
cryptography==2.1.4
cssselect==1.0.3
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
lxml==4.1.1
neo4j-driver==1.5.3
parsel==1.4.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.5.0
queuelib==1.4.2
recordtype==1.1
Scrapy==1.5.0
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
urlparse2==1.1.1
w3lib==1.19.0
zope.interface==4.4.3
